+++ ase/test/build_/test_supercells.py
@@ -1,8 +1,14 @@
+import itertools
+
 import numpy as np
 import pytest
 
 from ase.build import bulk
-from ase.build.supercells import make_supercell
+from ase.build.supercells import (
+    make_supercell,
+    get_deviation_from_optimal_cell_shape,
+    find_optimal_cell_shape,
+)
 
 
 @pytest.fixture()
@@ -10,20 +16,24 @@ def rng():
     return np.random.RandomState(seed=42)
 
 
-@pytest.fixture(params=[
-    bulk("NaCl", crystalstructure="rocksalt", a=4.0),
-    bulk("NaCl", crystalstructure="rocksalt", a=4.0, cubic=True),
-    bulk("Au", crystalstructure="fcc", a=4.0),
-])
+@pytest.fixture(
+    params=[
+        bulk("NaCl", crystalstructure="rocksalt", a=4.0),
+        bulk("NaCl", crystalstructure="rocksalt", a=4.0, cubic=True),
+        bulk("Au", crystalstructure="fcc", a=4.0),
+    ]
+)
 def prim(request):
     return request.param
 
 
-@pytest.fixture(params=[
-    3 * np.diag([1, 1, 1]),
-    4 * np.array([[1, 1, 0], [0, 1, 1], [1, 0, 1]]),
-    3 * np.diag([1, 2, 1]),
-])
+@pytest.fixture(
+    params=[
+        3 * np.diag([1, 1, 1]),
+        4 * np.array([[1, 1, 0], [0, 1, 1], [1, 0, 1]]),
+        3 * np.diag([1, 2, 1]),
+    ]
+)
 def P(request):
     return request.param
 
@@ -58,7 +68,7 @@ def test_make_supercells_arrays(prim, P, order, rng):
     assert reps * len(prim) == len(sc.get_tags())
     if order == "cell-major":
         assert all(sc.get_tags() == np.tile(tags, reps))
-        assert np.allclose(sc[:len(prim)].get_momenta(), prim.get_momenta())
+        assert np.allclose(sc[: len(prim)].get_momenta(), prim.get_momenta())
         assert np.allclose(sc.get_momenta(), np.tile(momenta, (reps, 1)))
     elif order == "atom-major":
         assert all(sc.get_tags() == np.repeat(tags, reps))
@@ -66,12 +76,15 @@ def test_make_supercells_arrays(prim, P, order, rng):
         assert np.allclose(sc.get_momenta(), np.repeat(momenta, reps, axis=0))
 
 
-@pytest.mark.parametrize('rep', [
-    (1, 1, 1),
-    (1, 2, 1),
-    (4, 5, 6),
-    (40, 19, 42),
-])
+@pytest.mark.parametrize(
+    "rep",
+    [
+        (1, 1, 1),
+        (1, 2, 1),
+        (4, 5, 6),
+        (40, 19, 42),
+    ],
+)
 def test_make_supercell_vs_repeat(prim, rep):
     P = np.diag(rep)
 
@@ -86,3 +99,59 @@ def test_make_supercell_vs_repeat(prim, rep):
     at2 = make_supercell(prim, P, wrap=False)
     assert np.allclose(at1.positions, at2.positions)
     assert all(at1.symbols == at2.symbols)
+
+
+def test_get_deviation_from_optimal_cell_shape():
+    # also tested via the docs data examples
+    # test perfect scores for SC, where cell vector permutation or magnitude
+    # do not matter:
+    cell = np.array([[1, 0, 0], [0, 1, 0], [0, 0, 1]])
+    for perm, factor in itertools.product(
+        itertools.permutations(range(3)), range(1, 9)
+    ):
+        permuted_cell = [cell[i] * factor for i in perm]
+        assert np.isclose(
+            get_deviation_from_optimal_cell_shape(permuted_cell, target_shape="sc"), 0.0
+        )
+
+    # likewise for FCC:
+    cell = np.array([[1, 1, 0], [0, 1, 1], [1, 0, 1]])
+    for perm, factor in itertools.product(itertools.permutations(range(3)),
+                                          range(1, 9)):
+        permuted_cell = [cell[i] * factor for i in perm]
+        assert np.isclose(
+            get_deviation_from_optimal_cell_shape(permuted_cell, target_shape="fcc"),
+            0.0,
+        )
+
+    # spot check some cases:
+    cell = np.array([[1, 0, 0], [0, 1, 0], [0, 0, 2]])
+    assert np.isclose(get_deviation_from_optimal_cell_shape(cell, "sc"), 0.6558650332)
+
+    # fcc
+    cell = np.array([[0, 1, 1], [1, 0, 1], [2, 2, 0]])
+    assert np.isclose(get_deviation_from_optimal_cell_shape(cell, "fcc"), 0.6558650332)
+
+    # negative determinant
+    cell = np.array([[-1, 0, 0], [0, -1, 0], [0, 0, -1]])
+    assert np.isclose(get_deviation_from_optimal_cell_shape(cell, "sc"), 0.0)
+
+
+def test_find_optimal_cell_shape():
+    # also tested via the docs data examples
+    cell = np.diag([1.0, 2.0, 4.0])
+    target_size = 8
+    target_shape = "sc"
+    result = find_optimal_cell_shape(cell, target_size, target_shape)
+    assert np.isclose(
+        get_deviation_from_optimal_cell_shape(np.dot(result, cell), "sc"), 0.0
+    )
+    assert np.allclose(np.linalg.norm(np.dot(result, cell), axis=1), 4)
+
+    # docs examples:
+    conf = bulk("Au")  # fcc
+    P1 = find_optimal_cell_shape(conf.cell, 32, "sc")
+    assert np.allclose(P1, np.array([[-2, 2, 2], [2, -2, 2], [2, 2, -2]]))
+
+    P1 = find_optimal_cell_shape(conf.cell, 495, "sc")
+    assert np.allclose(P1, np.array([[-6, 5, 5], [5, -6, 5], [5, 5, -5]]))